Update the floating precision to support double. Also update the output notation of numbers in console.

+++ Compiler/src/Number.cpp
@@ -1,7 +1,7 @@
 #include "pch.h"
 #include "Number.h"
 
-Number::Number(const std::variant<float, int, bool>& value) :
+Number::Number(const std::variant<double, int, bool>& value) :
     value(value),
     start(nullptr),
     end(nullptr) {
@@ -13,7 +13,7 @@ std::pair<Number*, Error*> Number::add(Number* other)
 	result->context = context;
     
     if (value.index() == 0 && other->value.index() == 0) {
-		try { result->value = std::get<float>(value) + std::get<float>(other->value); }
+		try { result->value = std::get<double>(value) + std::get<double>(other->value); }
 		catch (const std::bad_variant_access&) {}
     }
     else if(value.index() == 1 && other->value.index() == 1) {
@@ -21,11 +21,11 @@ std::pair<Number*, Error*> Number::add(Number* other)
 		catch (const std::bad_variant_access&) {}
     }
 	else if (value.index() == 0 && other->value.index() == 1) {
-		try { result->value = std::get<float>(value) + (float)std::get<int>(other->value); }
+		try { result->value = std::get<double>(value) + (double)std::get<int>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 0) {
-		try { result->value = (float)std::get<int>(value) + std::get<float>(other->value); }
+		try { result->value = (double)std::get<int>(value) + std::get<double>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 
@@ -38,7 +38,7 @@ std::pair<Number*, Error*> Number::subtract(Number* other)
 	result->context = context;
 
 	if (value.index() == 0 && other->value.index() == 0) {
-		try { result->value = std::get<float>(value) - std::get<float>(other->value); }
+		try { result->value = std::get<double>(value) - std::get<double>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 1) {
@@ -46,11 +46,11 @@ std::pair<Number*, Error*> Number::subtract(Number* other)
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 0 && other->value.index() == 1) {
-		try { result->value = std::get<float>(value) - (float)std::get<int>(other->value); }
+		try { result->value = std::get<double>(value) - (double)std::get<int>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 0) {
-		try { result->value = (float)std::get<int>(value) - std::get<float>(other->value); }
+		try { result->value = (double)std::get<int>(value) - std::get<double>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 
@@ -63,7 +63,7 @@ std::pair<Number*, Error*> Number::multiply(Number* other)
 	result->context = context;
 
 	if (value.index() == 0 && other->value.index() == 0) {
-		try { result->value = std::get<float>(value) * std::get<float>(other->value); }
+		try { result->value = std::get<double>(value) * std::get<double>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 1) {
@@ -71,11 +71,11 @@ std::pair<Number*, Error*> Number::multiply(Number* other)
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 0 && other->value.index() == 1) {
-		try { result->value = std::get<float>(value) * (float)std::get<int>(other->value); }
+		try { result->value = std::get<double>(value) * (double)std::get<int>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 0) {
-		try { result->value = (float)std::get<int>(value) * std::get<float>(other->value); }
+		try { result->value = (double)std::get<int>(value) * std::get<double>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 
@@ -104,12 +104,12 @@ std::pair<Number*, Error*> Number::divide(Number* other)
 
 	if (value.index() == 0 && other->value.index() == 0) {
 		try { 
-			auto rhs = std::get<float>(other->value);
+			auto rhs = std::get<double>(other->value);
 
-			if (rhs == 0.0f)
+			if (rhs == 0.0)
 				return std::make_pair(nullptr, new RuntimeError(other->start, other->end, "Division by zero", context));
 			else
-				result->value = std::get<float>(value) / rhs;
+				result->value = std::get<double>(value) / rhs;
 		}
 		catch (const std::bad_variant_access&) {}
 	}
@@ -120,7 +120,7 @@ std::pair<Number*, Error*> Number::divide(Number* other)
 			if (rhs == 0)
 				return std::make_pair(nullptr, new RuntimeError(other->start, other->end, "Division by zero", context));
 			else
-				result->value = (float)std::get<int>(value) / (float)rhs;
+				result->value = (double)std::get<int>(value) / (double)rhs;
 		}
 		catch (const std::bad_variant_access&) {}
 	}
@@ -131,18 +131,18 @@ std::pair<Number*, Error*> Number::divide(Number* other)
 			if (rhs == 0)
 				return std::make_pair(nullptr, new RuntimeError(other->start, other->end, "Division by zero", context));
 			else
-				result->value = std::get<float>(value) / (float)rhs;
+				result->value = std::get<double>(value) / (double)rhs;
 		}
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 0) {
 		try {
-			auto rhs = std::get<float>(other->value);
+			auto rhs = std::get<double>(other->value);
 
 			if (rhs == 0)
 				return std::make_pair(nullptr, new RuntimeError(other->start, other->end, "Division by zero", context));
 			else
-				result->value = (float)std::get<int>(value) / rhs;
+				result->value = (double)std::get<int>(value) / rhs;
 		}
 		catch (const std::bad_variant_access&) {}
 	}
@@ -156,19 +156,19 @@ std::pair<Number*, Error*> Number::power(Number* other)
 	result->context = context;
 
 	if (value.index() == 0 && other->value.index() == 0) {
-		try { result->value = (float)pow(std::get<float>(value), std::get<float>(other->value)); }
+		try { result->value = (double)pow(std::get<double>(value), std::get<double>(other->value)); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 1) {
-		try { result->value = (float)pow((float)std::get<int>(value), std::get<int>(other->value)); }
+		try { result->value = (double)pow((double)std::get<int>(value), std::get<int>(other->value)); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 0 && other->value.index() == 1) {
-		try { result->value = (float)pow(std::get<float>(value), (float)std::get<int>(other->value)); }
+		try { result->value = (double)pow(std::get<double>(value), (double)std::get<int>(other->value)); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 0) {
-		try { result->value = (float)pow((float)std::get<int>(value), std::get<float>(other->value)); }
+		try { result->value = (double)pow((double)std::get<int>(value), std::get<double>(other->value)); }
 		catch (const std::bad_variant_access&) {}
 	}
 
@@ -222,19 +222,19 @@ std::pair<Number*, Error*> Number::compare_and(Number* other)
 	Number* result = new Number();
 
 	if (value.index() == 0 && other->value.index() == 0) {
-		try { result->value = std::get<float>(value) && std::get<float>(other->value); }
+		try { result->value = std::get<double>(value) && std::get<double>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 1) {
-		try { result->value = (float)std::get<int>(value) && std::get<int>(other->value); }
+		try { result->value = (double)std::get<int>(value) && std::get<int>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 0 && other->value.index() == 1) {
-		try { result->value = std::get<float>(value) && (float)std::get<int>(other->value); }
+		try { result->value = std::get<double>(value) && (double)std::get<int>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 0) {
-		try { result->value = (float)std::get<int>(value) && std::get<float>(other->value); }
+		try { result->value = (double)std::get<int>(value) && std::get<double>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 2 && other->value.index() == 2) {
@@ -250,19 +250,19 @@ std::pair<Number*, Error*> Number::compare_or(Number* other)
 	Number* result = new Number();
 
 	if (value.index() == 0 && other->value.index() == 0) {
-		try { result->value = std::get<float>(value) || std::get<float>(other->value); }
+		try { result->value = std::get<double>(value) || std::get<double>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 1) {
-		try { result->value = (float)std::get<int>(value) || std::get<int>(other->value); }
+		try { result->value = (double)std::get<int>(value) || std::get<int>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 0 && other->value.index() == 1) {
-		try { result->value = std::get<float>(value) || (float)std::get<int>(other->value); }
+		try { result->value = std::get<double>(value) || (double)std::get<int>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1 && other->value.index() == 0) {
-		try { result->value = (float)std::get<int>(value) ||  std::get<float>(other->value); }
+		try { result->value = (double)std::get<int>(value) ||  std::get<double>(other->value); }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 2 && other->value.index() == 2) {
@@ -278,7 +278,7 @@ std::pair<Number*, Error*> Number::compare_not(Number* other)
 	Number* result = new Number();
 
 	if (value.index() == 0) {
-		try { result->value = std::get<float>(value) == 0.0f ? 1 : 0; }
+		try { result->value = std::get<double>(value) == 0.0 ? 1 : 0; }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1) {
@@ -294,7 +294,7 @@ std::pair<Number*, Error*> Number::is_true()
 	Number* result = new Number();
 
 	if (value.index() == 0) {
-		try { result->value = std::get<float>(value) != 0.0f; }
+		try { result->value = std::get<double>(value) != 0.0; }
 		catch (const std::bad_variant_access&) {}
 	}
 	else if (value.index() == 1) {
